+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # Hike ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Added the ability to deduplicate the history.
+  ([#105](https://github.com/davep/hike/pull/105))
+
 ## v0.12.0
 
 **Released: 2025-04-30**